Setup changes, Changes in the match details and scorecard

+++ lib/app/modules/SetupScreen/competition_player_model.dart
@@ -4,110 +4,118 @@
 
 import 'dart:convert';
 
-CompetitionPlayer competitionPlayerFromJson(String str) => CompetitionPlayer.fromJson(json.decode(str));
+CompetitionPlayer competitionPlayerFromJson(String str) =>
+    CompetitionPlayer.fromJson(json.decode(str));
 
-String competitionPlayerToJson(CompetitionPlayer data) => json.encode(data.toJson());
+String competitionPlayerToJson(CompetitionPlayer data) =>
+    json.encode(data.toJson());
 
 class CompetitionPlayer {
-    CompetitionPlayer({
-        this.status,
-        this.success,
-        this.message,
-        this.error,
-        this.payload,
-    });
+  CompetitionPlayer({
+    this.status,
+    this.success,
+    this.message,
+    this.error,
+    this.payload,
+  });
 
-    int? status;
-    bool? success;
-    String? message;
-    String? error;
-    List<Payload>? payload;
+  int? status;
+  bool? success;
+  String? message;
+  String? error;
+  List<Payload>? payload;
 
-    factory CompetitionPlayer.fromJson(Map<String, dynamic> json) => CompetitionPlayer(
+  factory CompetitionPlayer.fromJson(Map<String, dynamic> json) =>
+      CompetitionPlayer(
         status: json["status"],
         success: json["success"],
         message: json["message"],
         error: json["error"],
-        payload: List<Payload>.from(json["payload"].map((x) => Payload.fromJson(x))),
-    );
+        payload:
+            List<Payload>.from(json["payload"].map((x) => Payload.fromJson(x))),
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "status": status,
         "success": success,
         "message": message,
         "error": error,
         "payload": List<dynamic>.from(payload!.map((x) => x.toJson())),
-    };
+      };
 }
 
 class Payload {
-    Payload({
-        this.id,
-        this.compTime,
-        this.compHcp,
-        this.lastPlayedHole,
-        this.totalResult,
-        this.player,
-        this.recordingScoresFor,
-        this.scorecards,
-    });
+  Payload({
+    this.id,
+    this.compTime,
+    this.compHcp,
+    this.lastPlayedHole,
+    this.totalResult,
+    this.player,
+    this.recordingScoresFor,
+    this.scorecards,
+  });
 
-    int? id;
-    String? compTime;
-    int? compHcp;
-    LastPlayedHole? lastPlayedHole;
-    int? totalResult;
-    Player? player;
-    Player? recordingScoresFor;
-    List<Scorecard>? scorecards;
+  int? id;
+  String? compTime;
+  int? compHcp;
+  LastPlayedHole? lastPlayedHole;
+  double? totalResult;
+  Player? player;
+  Player? recordingScoresFor;
+  List<Scorecard>? scorecards;
 
-    factory Payload.fromJson(Map<String, dynamic> json) => Payload(
+  factory Payload.fromJson(Map<String, dynamic> json) => Payload(
         id: json["id"],
         compTime: json["compTime"],
         compHcp: json["compHcp"],
-        lastPlayedHole: json["lastPlayedHole"] == null ? null : LastPlayedHole.fromJson(json["lastPlayedHole"]),
+        lastPlayedHole: json["lastPlayedHole"] == null
+            ? null
+            : LastPlayedHole.fromJson(json["lastPlayedHole"]),
         totalResult: json["totalResult"],
         player: Player.fromJson(json["player"]),
         recordingScoresFor: Player.fromJson(json["recordingScoresFor"]),
-        scorecards: List<Scorecard>.from(json["scorecards"].map((x) => Scorecard.fromJson(x))),
-    );
+        scorecards: List<Scorecard>.from(
+            json["scorecards"].map((x) => Scorecard.fromJson(x))),
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "id": id,
         "compTime": compTime,
         "compHcp": compHcp,
-        "lastPlayedHole": lastPlayedHole == null ? null : lastPlayedHole!.toJson(),
+        "lastPlayedHole":
+            lastPlayedHole == null ? null : lastPlayedHole!.toJson(),
         "totalResult": totalResult,
         "player": player!.toJson(),
         "recordingScoresFor": recordingScoresFor!.toJson(),
         "scorecards": List<dynamic>.from(scorecards!.map((x) => x.toJson())),
-    };
+      };
 }
 
 class LastPlayedHole {
-    LastPlayedHole({
-        this.id,
-        this.holeNo,
-        this.yellow,
-        this.white,
-        this.blue,
-        this.red,
-        this.par,
-        this.stroke,
-        this.courseId,
-    });
+  LastPlayedHole({
+    this.id,
+    this.holeNo,
+    this.yellow,
+    this.white,
+    this.blue,
+    this.red,
+    this.par,
+    this.stroke,
+    this.courseId,
+  });
 
-    int? id;
-    int? holeNo;
-    int? yellow;
-    int? white;
-    int? blue;
-    int? red;
-    int? par;
-    int? stroke;
-    int? courseId;
+  int? id;
+  int? holeNo;
+  int? yellow;
+  int? white;
+  int? blue;
+  int? red;
+  int? par;
+  int? stroke;
+  int? courseId;
 
-    factory LastPlayedHole.fromJson(Map<String, dynamic> json) => LastPlayedHole(
+  factory LastPlayedHole.fromJson(Map<String, dynamic> json) => LastPlayedHole(
         id: json["id"],
         holeNo: json["holeNo"],
         yellow: json["yellow"],
@@ -117,9 +125,9 @@ class LastPlayedHole {
         par: json["par"],
         stroke: json["stroke"],
         courseId: json["courseId"],
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "id": id,
         "holeNo": holeNo,
         "yellow": yellow,
@@ -129,37 +137,37 @@ class LastPlayedHole {
         "par": par,
         "stroke": stroke,
         "courseId": courseId,
-    };
+      };
 }
 
 class Player {
-    Player({
-        this.id,
-        this.fname,
-        this.lname,
-        this.address,
-        this.dob,
-        this.gender,
-        this.dateJoined,
-        this.image,
-        this.hcp,
-        this.fcmToken,
-        this.aspNetUsers,
-    });
+  Player({
+    this.id,
+    this.fname,
+    this.lname,
+    this.address,
+    this.dob,
+    this.gender,
+    this.dateJoined,
+    this.image,
+    this.hcp,
+    this.fcmToken,
+    this.aspNetUsers,
+  });
 
-    int? id;
-    String? fname;
-    String? lname;
-    String? address;
-    DateTime? dob;
-    String? gender;
-    DateTime? dateJoined;
-    String? image;
-    int? hcp;
-    String? fcmToken;
-    AspNetUsers? aspNetUsers;
+  int? id;
+  String? fname;
+  String? lname;
+  String? address;
+  DateTime? dob;
+  String? gender;
+  DateTime? dateJoined;
+  String? image;
+  int? hcp;
+  String? fcmToken;
+  AspNetUsers? aspNetUsers;
 
-    factory Player.fromJson(Map<String, dynamic> json) => Player(
+  factory Player.fromJson(Map<String, dynamic> json) => Player(
         id: json["id"],
         fname: json["fname"],
         lname: json["lname"],
@@ -171,9 +179,9 @@ class Player {
         hcp: json["hcp"],
         fcmToken: json["fcmToken"],
         aspNetUsers: AspNetUsers.fromJson(json["aspNetUsers"]),
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "id": id,
         "fname": fname,
         "lname": lname,
@@ -185,69 +193,69 @@ class Player {
         "hcp": hcp,
         "fcmToken": fcmToken,
         "aspNetUsers": aspNetUsers!.toJson(),
-    };
+      };
 }
 
 class AspNetUsers {
-    AspNetUsers({
-        this.id,
-        this.userName,
-        this.email,
-        this.phoneNumber,
-    });
+  AspNetUsers({
+    this.id,
+    this.userName,
+    this.email,
+    this.phoneNumber,
+  });
 
-    String? id;
-    String? userName;
-    String? email;
-    dynamic phoneNumber;
+  String? id;
+  String? userName;
+  String? email;
+  dynamic phoneNumber;
 
-    factory AspNetUsers.fromJson(Map<String, dynamic> json) => AspNetUsers(
+  factory AspNetUsers.fromJson(Map<String, dynamic> json) => AspNetUsers(
         id: json["id"],
         userName: json["userName"],
         email: json["email"],
         phoneNumber: json["phoneNumber"],
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "id": id,
         "userName": userName,
         "email": email,
         "phoneNumber": phoneNumber,
-    };
+      };
 }
 
 class Scorecard {
-    Scorecard({
-        this.id,
-        this.score,
-        this.timestamp,
-        this.confirmed,
-        this.result,
-        this.holeNo,
-    });
+  Scorecard({
+    this.id,
+    this.score,
+    this.timestamp,
+    this.confirmed,
+    this.result,
+    this.holeNo,
+  });
 
-    int? id;
-    int? score;
-    String? timestamp;
-    bool? confirmed;
-    int? result;
-    int? holeNo;
+  int? id;
+  int? score;
+  String? timestamp;
+  bool? confirmed;
+  int? result;
+  int? holeNo;
 
-    factory Scorecard.fromJson(Map<String, dynamic> json) => Scorecard(
+  factory Scorecard.fromJson(Map<String, dynamic> json) => Scorecard(
         id: json["id"],
         score: json["score"],
         timestamp: json["timestamp"],
         confirmed: json["confirmed"],
         result: json["result"],
         holeNo: json["holeNo"],
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "id": id,
         "score": score,
         "timestamp": timestamp,
         "confirmed": confirmed,
         "result": result,
         "holeNo": holeNo,
-    };
+      };
 }
